wrapped sdk init outside specific module
