+++ android/src/main/kotlin/br/com/adriankohls/zendesk2/Zendesk2Chat.kt
@@ -86,13 +86,6 @@ class Zendesk2Chat(private val activity: Activity?, private val channel: MethodC
                 else -> PreChatFormFieldStatus.HIDDEN
             }
 
-    fun init(call: MethodCall) {
-        val accountKey = call.argument<String>("accountKey")!!
-        val appId = call.argument<String>("appId")!!
-
-        Chat.INSTANCE.init(activity!!, accountKey, appId)
-    }
-
     fun logger(call: MethodCall) {
         var enabled = call.argument<Boolean>("enabled")
         enabled = enabled ?: false
